+++ src/os2datascanner/projects/admin/import_services/models/ldap_import_job.py
@@ -3,6 +3,7 @@ from django.db import models
 from os2datascanner.utils.ldap import RDN
 
 from os2datascanner.projects.admin.adminapp.signals import get_pika_thread
+from os2datascanner.projects.admin.adminapp.utils import sync_all_accounts_and_scanners
 from ...core.models.background_job import BackgroundJob
 from .realm import Realm
 
@@ -50,6 +51,8 @@ class LDAPImportJob(BackgroundJob):
                 self.save()
 
         perform_import(self.realm, progress_callback=_callback)
+
+        sync_all_accounts_and_scanners()
 
     def finish(self):
         if (pe := get_pika_thread(init=False)):